CG working for the non obstacle case (#51)
